requirements.txt
fastapi
uvicorn
python-multipart
ffmpeg-python
openai-whisper
opencv-python

pytest
pytest-asyncio
httpx